requirements.txt
numpy
pillow
matplotlib
scikit-image